handle delete room with associated images

+++ src/app/api/admin/rooms/route.ts
@@ -1,8 +1,9 @@
 import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
 import { OPTIONS } from "../../auth/[...nextauth]/route";
-import Room from "@/models/roomModel";
+import Room, { RoomDocument } from "@/models/roomModel";
 import { revalidatePath } from "next/cache";
+import { deleteImage } from "@/lib/cloudinary";
 
 export async function GET(req: NextRequest) {
   try {
@@ -26,17 +27,27 @@ export async function DELETE(req: NextRequest) {
   try {
     const session = await getServerSession(OPTIONS);
     if (session?.user?.role !== "admin") {
-      throw new Error("Only Admin Can Show This Route");
+      throw new Error("Only Admin Can Delete Room");
     }
     const roomId = req.nextUrl.searchParams.get("roomId");
 
+    const room: RoomDocument | null = await Room.findById(roomId);
+
+    if (!room) {
+      throw new Error("No Room Found");
+    }
+
+    if (room?.images) {
+      room.images.map(async (img) => await deleteImage(img.public_id));
+    }
+
     await Room.findByIdAndDelete(roomId);
 
-    revalidatePath(`${process.env.SITE_URL}/admin/rooms`);
+    revalidatePath("/admin/rooms");
 
     return NextResponse.json(
       { message: "Room Deleted Successfully" },
-      { status: 200 }
+      { status: 201 }
     );
   } catch (error: any) {
     return NextResponse.json({ message: error.message }, { status: 500 });
